Testing of ExpenseListFilters methods
Props had to be passed into one of the methods so a new fixture was made and setProps was used here to render new prop data.

Most methods required an event handler target value to be passed along with the change and this had to be included as a 2nd argument in the simulate function as an object, containing a key: object(value).

The 3rd party dateRangePicker had a method prop that .prop accessed and tested for a value change using a state selector in the expect test method.