Setup production build and server
Creating a port that heroku can supply with or that defaults to 3000 if not supplied.

There was an issue regarding development VS production enviroments not having the right assets on push by nature of how these bundled files are being made - a .gitignore was added pointing to these bundle files to prevent this issue.